Major - PutOnSale works!
Done PutOnSale module. Bot could set items for sale, check if there're
extra or not enough items on sale and correct it.
Created the StopSellingItem and RenewPriceOnSellingItem functions in
Csgotm api. Bot could correct price for given itemId or delete it.
Little bit cleaned up the code. For now it's fully working bot.
In plans - test the bot with ~20 items in database. Create GUI for it.
Clean up the code. For now it's a real mess.